+++ tests/fake_smbus2.py
@@ -4,11 +4,10 @@ class SMBus():
         self.bus_number = bus_number
         self.fake_register_output = 0
 
-    def i2c_rdwr(self, write, read=None):
-        self.fake_register_input = write
-        if read is not None:
-            read.clear()
-            read.extend(self.fake_register_output)
+    def i2c_rdwr(self, operation):
+        self.fake_register_input = operation
+        if operation == [14, 1]:
+            operation[0] = self.fake_register_output
 
     def fakeInput(self):
         return self.fake_register_input